[Refactor] Book도메인과 Aladin도메인 분리
