+++ src/main/java/com/example/bookiibookii/domain/aladin/exception/AladinException.java
@@ -1,0 +1,10 @@
+package com.example.bookiibookii.domain.aladin.exception;
+
+import com.example.bookiibookii.global.apiPayload.code.BaseCode;
+import com.example.bookiibookii.global.apiPayload.exception.GeneralException;
+
+public class AladinException extends GeneralException {
+    public AladinException(BaseCode code) {
+        super(code);
+    }
+}
